* recover-src/extract_list.c (delete_file): j is int. * changer-src/scsi-defs.h (B, B1): Check if already defined. * server-src/taper.c: Some cast to int. * tape-src/amdd.c (read_func, write_func): Complete prototype. * tape-src/tapeio.h (tapefd_read, tapefd_write): return ssize_t,   third argument is size_t. * tape-src/tapeio.c: Ditto. * tape-src/output-file.h (file_tapefd_read, file_tapefd_write): return   ssize_t, third argument is size_t. * tape-src/output-file.c: Ditto. * tape-src/output-null.h (null_tapefd_read, null_tapefd_write): return   ssize_t, third argument is size_t. * tape-src/output-null.c: Ditto. * tape-src/output-rait.h (rait_read, rait__write): return ssize_t,   third argument is size_t. * tape-src/output-rait.c: Ditto. * tape-src/output-tape.h (tape_tapefd_read, tape_tapefd_write): return   ssize_t, third argument is size_t. * tape-src/output-tape.c: Ditto.

+++ recover-src/extract_list.c
@@ -24,7 +24,7 @@
  * file named AUTHORS, in the root directory of this distribution.
  */
 /*
- * $Id: extract_list.c,v 1.43.2.13.4.6.2.12 2002/11/12 19:19:03 martinea Exp $
+ * $Id: extract_list.c,v 1.43.2.13.4.6.2.13 2002/11/12 21:24:20 martinea Exp $
  *
  * implements the "extract" command in amrecover
  */
@@ -704,7 +704,7 @@ char *regex;
     char *cmd = NULL;
     char *err = NULL;
     int i;
-    size_t j;
+    int j;
     char *date, *date_undo, date_undo_ch = '\0';
     char *tape, *tape_undo, tape_undo_ch = '\0';
     char *dir, *dir_undo, dir_undo_ch = '\0';
